Rename file. Pull answers an photos from db in left join
